+++ packages/showcases/svelte/stories/components/VtmnButton/VtmnButton.stories.js
@@ -1,4 +1,5 @@
 import VtmnButtonView from './VtmnButtonView.svelte';
+import vitamixIconsList from '@vtmn/icons/dist/vitamix/font/vitamix.json';
 import VtmnButton from '@vtmn/svelte/src/components/VtmnButton.svelte';
 
 export default {
@@ -28,6 +29,33 @@ export default {
       control: {
         type: 'select',
         options: ['small', 'medium', 'large', 'stretched'],
+      },
+    },
+    iconLeft: {
+      type: { name: 'string', required: false },
+      description: 'The icon to display on the left hand side of button.',
+      defaultValue: null,
+      control: {
+        type: 'select',
+        options: ['', ...Object.keys(vitamixIconsList)],
+      },
+    },
+    iconRight: {
+      type: { name: 'string', required: false },
+      description: 'The icon to display on the right hand side of button.',
+      defaultValue: null,
+      control: {
+        type: 'select',
+        options: ['', ...Object.keys(vitamixIconsList)],
+      },
+    },
+    iconAlone: {
+      type: { name: 'string', required: false },
+      description: 'The to display when it is a button with icon only.',
+      defaultValue: null,
+      control: {
+        type: 'select',
+        options: ['', ...Object.keys(vitamixIconsList)],
       },
     },
   },