+++ packages/sources/react/src/index.ts
@@ -1,7 +1,9 @@
 import '@vtmn/css-design-tokens/dist/index.css';
 
 export { VtmnButton } from './components/VtmnButton';
+export { VtmnCheckbox } from './components/VtmnCheckbox';
 export { VtmnIcon } from './components/VtmnIcon';
 export { VtmnLink } from './components/VtmnLink';
+export { VtmnRadioButton } from './components/VtmnRadioButton';
 export { VtmnTextInput } from './components/VtmnTextInput';
 export { VtmnToggle } from './components/VtmnToggle';