+++ packages/sources/svelte/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.4.1...@vtmn/svelte@0.4.2) (2021-06-24)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+## [0.4.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.4.0...@vtmn/svelte@0.4.1) (2021-06-24)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+# [0.4.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.3.13...@vtmn/svelte@0.4.0) (2021-06-23)
+
+
+### Features
+
+* **@vtmn/svelte:** add icons in button component ([#381](https://github.com/Decathlon/vitamin-web/issues/381)) ([2a77b91](https://github.com/Decathlon/vitamin-web/commit/2a77b9155db257212b5e5dc5a73b0cba6f1143bf))
+
+
+
+
+
+## [0.3.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.3.12...@vtmn/svelte@0.3.13) (2021-06-21)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+## [0.3.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.3.11...@vtmn/svelte@0.3.12) (2021-06-21)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+## [0.3.11](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.3.10...@vtmn/svelte@0.3.11) (2021-06-17)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+## [0.3.10](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.3.9...@vtmn/svelte@0.3.10) (2021-06-17)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
 ## [0.3.9](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.3.8...@vtmn/svelte@0.3.9) (2021-06-14)
 
 **Note:** Version bump only for package @vtmn/svelte