+++ packages/sources/vue/CHANGELOG.md
@@ -3,6 +3,112 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.6.1...@vtmn/vue@0.6.2) (2021-06-25)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+## [0.6.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.6.0...@vtmn/vue@0.6.1) (2021-06-24)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+# [0.6.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.5.0...@vtmn/vue@0.6.0) (2021-06-24)
+
+
+### Features
+
+* **@vtmn/vue:** add toggle component ([#383](https://github.com/Decathlon/vitamin-web/issues/383)) ([027fb90](https://github.com/Decathlon/vitamin-web/commit/027fb90a11a86cb46e7160a59df1cae98490f992))
+
+
+
+
+
+# [0.5.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.4.0...@vtmn/vue@0.5.0) (2021-06-23)
+
+
+### Features
+
+* **@vtmn/vue:** add icons in button component ([#382](https://github.com/Decathlon/vitamin-web/issues/382)) ([6ea5eb8](https://github.com/Decathlon/vitamin-web/commit/6ea5eb8f009e48edb368fbc278f1a5c2de0e748d))
+
+
+
+
+
+# [0.4.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.3.0...@vtmn/vue@0.4.0) (2021-06-21)
+
+
+### Features
+
+* **@vtmn/vue:** add text input component ([#359](https://github.com/Decathlon/vitamin-web/issues/359)) ([0589891](https://github.com/Decathlon/vitamin-web/commit/058989123c0282b126fddb7021d887a05f94ab8b))
+
+
+
+
+
+# [0.3.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.2.1...@vtmn/vue@0.3.0) (2021-06-21)
+
+
+### Features
+
+* **@vtmn/vue:** add link component ([#358](https://github.com/Decathlon/vitamin-web/issues/358)) ([7de0dd7](https://github.com/Decathlon/vitamin-web/commit/7de0dd77ca00686fabbd1fec1f2db6f75bef7068)), closes [#341](https://github.com/Decathlon/vitamin-web/issues/341)
+
+
+
+
+
+## [0.2.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.2.0...@vtmn/vue@0.2.1) (2021-06-17)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+# [0.2.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.1.0...@vtmn/vue@0.2.0) (2021-06-17)
+
+
+### Features
+
+* **@vtmn/vue:** add checkbox component ([#363](https://github.com/Decathlon/vitamin-web/issues/363)) ([b3258c4](https://github.com/Decathlon/vitamin-web/commit/b3258c4374dac5a22d3e208cb785b444b6bdfdfd))
+
+
+
+
+
+# [0.1.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.0.31...@vtmn/vue@0.1.0) (2021-06-17)
+
+
+### Features
+
+* **@vtmn/vue:** add radio button component ([#360](https://github.com/Decathlon/vitamin-web/issues/360)) ([96f7c48](https://github.com/Decathlon/vitamin-web/commit/96f7c48fd978a85927d6d82ad8ee5c7dddd1608a))
+
+
+
+
+
+## [0.0.31](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.0.30...@vtmn/vue@0.0.31) (2021-06-17)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+## [0.0.30](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.0.29...@vtmn/vue@0.0.30) (2021-06-17)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
 ## [0.0.29](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.0.28...@vtmn/vue@0.0.29) (2021-06-14)
 
 **Note:** Version bump only for package @vtmn/vue